refactory validations instance
